split out mixins in views
